// src/generated/bindings/FfmpegTask.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FfmpegTaskStatus } from "./FfmpegTaskStatus";
import type { FfmpegTaskType } from "./FfmpegTaskType";

export type FfmpegTask = { status: FfmpegTaskStatus, taskType: FfmpegTaskType, };